third_party/WebKit/LayoutTests/netinfo/multiple-frames-expected.txt

Tests using NetInfo from multiple frames.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS childConnection.type is newConnectionType
PASS connection.type is newConnectionType
PASS successfullyParsed is true

TEST COMPLETE
